code/info_bar_app/requirements.txt

Flask==1.1.2
Flask_Migrate==3.1.0
Flask_SQLAlchemy==2.5.1
newspaper3k
numpy==1.20.3
scikit_learn==1.0.2
spacy>=2.0.0,<3.0.0
SQLAlchemy==1.3.23
textblob==0.17.1
vaderSentiment==3.3.2
gunicorn
markupsafe==2.0.1
itsdangerous==1.1.0
Jinja2==2.11.2
nltk
psycopg2-binary
